Move workflow test utils to the workflows app
